Fix path for `history.py`
Before path was "~/.local/share", which is a fallback path for profanity.
Profanity tries to get path from XDG_DATA_HOME variable
as "home" and then make a profanity folder there
with logs and other things,
and only then falls back to "~/.local/share",
but the plugin was using the second path as default.